Move AbiMetadata from include/lgc/state to interface/lgc
This is because ABI metadata will be accessed by front-end in the future
to extract certain data and perform ELF patching. Current directory
hierarchy disallows us to do so.
